+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2026-06-19]
+### Added
+- Added support for HTLF2-Claymore Unit type.
+
 ## [2026-06-07]
 ### Added
 - Added support for toolhead sensor runout for standalone toolheads for toolchangers
@@ -67,6 +71,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2026-04-04]
 ### Fixed
 - Fixed issue where lane was trying to be looked up by keyname with `self.current` when the property returned `None`. Switched to using `self.lanes.get` since this is a safer operation.
+
+## [2026-04-01]
+### Added
+- Added support in the `install-afc.sh` script for the HTLF Claymore.
 
 ## [2026-03-30]
 ### Fixed